Changed the order of walls in the level format
Changed how the walls are loaded in the editor scene so that they can be loaded in the same way in this program and in the game (there was an unintended side effect of how levels loaded where the order of walls would change between different sessions based on the order the engine loaded them, oops).

How this works now is that the walls are stored in two arrays (I was gonna go with a matrix but that can't be done in the inspector for some reason) and are forced into a specific order set within the inspector. This prevents the walls from changing order between sessions.

The addition of the level controller script and the level scene are leftovers from when this was going to be for the game itself and not just the design tool.