Added TMPro and began par implementation
Added the files for TMPro and began to create the par feature.

Right now, the par feature is a textbox and slider that can take numbers between 0 and 99 inclusive. This is not yet implemented into the save data yet though.

TMPro is currently used to give the user a better sense of what edit mode they're in.

Also added two edit mode buttons because I keep forgetting my own key commands lol.